Refact code and add PlayerController
