--- Assets/Simple First Person Controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0545f7d3f50e1a546a565ab207d6e6c5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
