--- Assets/Simple First Person Controller/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6f7ccc9517ccda44cb8fec241c9e8a89
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
