--- Assets/Simple First Person Controller/Resources/Black.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e14942a3b8d9aee40a5ba3869d1ad2a3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
