--- Assets/Simple First Person Controller/Resources/Red.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c0772b21f0f4f5d4fa3e104d18061880
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
